Seperating each compoment from FilterableProductTable.jsx
